grt: fix bad min compare in FastRouteCore::netpinOrderInc
< should have been >, replaced by std::min

    int xmin = BIG_INT;
...
      if (xmin < treenodes[stree->edges[ind].n1].x) {
        xmin = treenodes[stree->edges[ind].n1].x;
      }

+++ src/grt/src/fastroute/include/DataType.h
@@ -213,7 +213,7 @@ struct OrderTree
 {
   int length;
   int treeIndex;
-  int xmin;
+  int16_t xmin;
 };
 
 struct OrderNetEdge
